metadata_manager: move VersionsFetcher and APSourceMetadataFetcher into separate files
