+++ metadata_manager/__init__.py
@@ -1,17 +1,14 @@
-from .core import (
-    APSourceMetadataFetcher,
+from .versions_fetcher import (
     VersionsFetcher,
     RemoteInfo,
 )
-from .exceptions import (
-    MetadataManagerException,
-    TooManyInstancesError
+
+from .ap_src_meta_fetcher import (
+    APSourceMetadataFetcher,
 )
 
 __all__ = [
     "APSourceMetadataFetcher",
     "VersionsFetcher",
-    "MetadataManagerException",
-    "TooManyInstancesError",
     "RemoteInfo",
 ]
